requirements.txt
mutagen
pyacoustid
musicbrainzngs
fastapi
uvicorn
python-multipart
aiofiles
requests
chromaprint
pymongo
